Refactor partial file layout for saves and decoders
