+++ CodeGlyphX/BarcodeDecoder.Decoders.EanUpc.cs
