+++ CodeGlyphX/BarcodeDecoder.Decoders.Linear.cs
