+++ CodeGlyphX/DataMatrix/DataMatrixDecoder.Internal.Encodations.cs
