+++ CodeGlyphX/DataMatrix/DataMatrixDecoder.Internal.Pipeline.cs
