+++ CodeGlyphX/QrEasy.Render.Core.cs
