+++ CodeGlyphX/QrEasy.Render.Extras.cs
