+++ CodeGlyphX/Rendering/Jpeg/JpegReader.Decode.Helpers.cs
