+++ CodeGlyphX/Rendering/Png/QrPngRenderer.Internal.Eyes.cs
