+++ CodeGlyphX/Rendering/Png/QrPngRenderer.Internal.Render.cs
